refactor(blockchain): make testchain its own module
